Réorganisation du code, mise en cohérence
- Amélioration des boutons graphiques
- Déplacement de certaines fonctions entre les fichiers
- Mise sous commentaire de toutes les fonctions et prototypes de NetworkArray
- Ajout d'un fichier "tools" qui rassemble des fonctions diverses
