+++ NetworkArray.h
@@ -7,97 +7,97 @@
 
 class Network;
 
-//! Classe comprenant tous les réseaux avec des méthodes groupées.
-class NetworkArray {
-public:
-
-//! Constructeur : crée les réseaux, et charge les options.
-NetworkArray(int length_alphabet = LENGTH_ALPHABET);
-
-//! Destructeur : détruit les réseaux.
-~NetworkArray();
-
-//METHODES
-
-//! Teste l'entree sur les reseaux.
-/*!	Fait passer l'entree dans les reseaux et retourne la lettre du reseau maximal.
-    \param input entree du reseau
-    \param verbose bool (optionnal : default = true ) si true : affiche les infos de sortie
-    \return la sortie la plus probable, ou '_' si aucun reseau n'a plus de LOWER_BOUND_DISTINCTION
- */
-char testNetworks(double input[], bool verbose = true);
-
-//! Teste toutes les entrees du repertoire de teste
-/*! Fait passer toutes les entrees dans les reseaux
-    \param directory string (optionnal : default = DOSSIERTESTTEXT ) dossier de test
-    \return stat double ratio of succes
- */
-double testAll(std::string directory = DOSSIERTESTTEXT);
-
-//! Lance l'apprentissage des reseaux.
-/*!	Chaque reseau apprend l'un après l'autre sur les exemples donnes dans DOSSIERTEXTES,
-    l'apprentissage d'un reseau s'arrete quand sa distance_maximale sur les exemples est inferieure à maximal_distance
-    ou quand il a deja effectue max_limit_loop apprentissages .
- */
-void learnAllNetworks();
-
-//! Actualise les reseaux.
-/*!
-    Chaque reseau reprend les poids enregistres dans le fichier specifie dans mostRecent.txt
- */
-void getMostRecent();
-
-//! Actualise les options des reseaux.
-void setOptions();
-
-//! Retourne le moment d'inertie des reseaux.
-/*!
-   \return momentum_ : le moment d'inertie du reseau */
-double getMomentum();
-
-//! Fixe le moment d'inertie des reseaux
-/*! selon le moment :
-   \param momentum nouveau moment d'inertie
- */
-void setMomentum(double momentum);
-
-//! Retourne l'alphabet des reseaux
-char* getAlphabet();
-
-//! Fixe des reseaux
-void setAlphabet(char* alphabet);
-
-//! Retourne la distance maximale apres apprentissage productif
-double getMaximalDistance();
-
-//! Fixe la distance maximale apres apprentissage productif des reseaux
-void setMaximalDistance(double maximal_distance);
-
-//! Retourne le nombre maximal de boucles d'apprentissage a effectuer a chaque apprentissage
-int  getMaxLimitLoop();
-
-//! Fixe le nombre maximal de boucles d'apprentissage a effectuer a chaque apprentissage
-void setMaxLimitLoop(int maxLimitLoop);
-
-void save();
-
-private:
-//! Le tableau des networks
-Network** tablo_net_;
-//! La distance maximale après un apprentissage producti
-double maximal_distance_;
-//! La limite du nombre de boucles d'apprentissag
-int maxLimitLoop_;
-//! L'alphabet concerne
-const char* m_alphabet;
-//! La longueur de l'alphabet concern
-int length_alphabet_;
-//! Le moment d'inertie de tous les réseau
-double momentum_;
-};
+////! Classe comprenant tous les réseaux avec des méthodes groupées.
+//class NetworkArray {
+//public:
+//
+////! Constructeur : crée les réseaux, et charge les options.
+//NetworkArray(int length_alphabet = LENGTH_ALPHABET);
+//
+////! Destructeur : détruit les réseaux.
+//~NetworkArray();
+//
+////METHODES
+//
+////! Teste l'entree sur les reseaux.
+///*!	Fait passer l'entree dans les reseaux et retourne la lettre du reseau maximal.
+//    \param input entree du reseau
+//    \param verbose bool (optionnal : default = true ) si true : affiche les infos de sortie
+//    \return la sortie la plus probable, ou '_' si aucun reseau n'a plus de LOWER_BOUND_DISTINCTION
+// */
+//char testNetworks(double input[], bool verbose = true);
+//
+////! Teste toutes les entrees du repertoire de teste
+///* Fait passer toutes les entrees dans les reseaux
+//    \param directory string (optionnal : default = DOSSIERTESTTEXT ) dossier de test
+//    \return stat double ratio of succes
+// */
+//double testAll(std::string directory = DOSSIERTESTTEXT);
+//
+////! Lance l'apprentissage des reseaux.
+///*!	Chaque reseau apprend l'un après l'autre sur les exemples donnes dans DOSSIERTEXTES,
+//    l'apprentissage d'un reseau s'arrete quand sa distance_maximale sur les exemples est inferieure à maximal_distance
+//    ou quand il a deja effectue max_limit_loop apprentissages .
+// */
+////void learnAllNetworks();
+//
+////! Actualise les reseaux.
+///*!
+//    Chaque reseau reprend les poids enregistres dans le fichier specifie dans mostRecent.txt
+// */
+////void getMostRecent();
+//
+////! Actualise les options des reseaux.
+//void setOptions();
+//
+////! Retourne le moment d'inertie des reseaux.
+///*!
+//   \return momentum_ : le moment d'inertie du reseau */
+//double getMomentum();
+//
+////! Fixe le moment d'inertie des reseaux
+///*! selon le moment :
+//   \param momentum nouveau moment d'inertie
+// */
+//void setMomentum(double momentum);
+//
+////! Retourne l'alphabet des reseaux
+//char* getAlphabet();
+//
+////! Fixe des reseaux
+//void setAlphabet(char* alphabet);
+//
+////! Retourne la distance maximale apres apprentissage productif
+//double getMaximalDistance();
+//
+////! Fixe la distance maximale apres apprentissage productif des reseaux
+//void setMaximalDistance(double maximal_distance);
+//
+////! Retourne le nombre maximal de boucles d'apprentissage a effectuer a chaque apprentissage
+//int  getMaxLimitLoop();
+//
+////! Fixe le nombre maximal de boucles d'apprentissage a effectuer a chaque apprentissage
+//void setMaxLimitLoop(int maxLimitLoop);
+//
+////void save();
+//
+//private:
+////! Le tableau des networks
+//Network** tablo_net_;
+////! La distance maximale après un apprentissage producti
+//double maximal_distance_;
+////! La limite du nombre de boucles d'apprentissag
+//int maxLimitLoop_;
+////! L'alphabet concerne
+//const char* m_alphabet;
+////! La longueur de l'alphabet concern
+//int length_alphabet_;
+////! Le moment d'inertie de tous les réseau
+//double momentum_;
+//};
 
 //! affichage d'un tableau
-template <class T>
-void displayArray(T* data, int length);
+//template <class T>
+//void displayArray(T* data, int length);
 
 #endif  //NETWORKARRAY_H_INCLUDED
