+++ script.cpp
@@ -36,7 +36,7 @@ void commands(int nbCmds, string cmds[])
 
 	getConfigValue(&length_alphabet, &mu, &maximal_distance, &momentum, &layerNb, sizesBis);
 	int sizes[layerNb];
-	for (size_t i = 0; i < layerNb; i++)
+	for (int i = 0; i < layerNb; i++)
 		sizes[i] = sizesBis[i];
 
 	ReadNetwork* rdnk = load(string(DOSSIERBACKUP) + string(NOMBACKUP), false);
@@ -103,7 +103,7 @@ void commands(int nbCmds, string cmds[])
 				}
 				layerNb = index+1;
 				int sizes[layerNb];
-				for (size_t j = 0; j < layerNb; j++)
+				for (int j = 0; j < layerNb; j++)
 					sizes[j] = sizesbis[j];
 			}
 		}
@@ -131,7 +131,7 @@ void commands(int nbCmds, string cmds[])
 		delete rdnk;
 }
 
-int* getConfigValue(int* length_alphabet, double* mu, double* maximal_distance, double* momentum, int* layerNb, int sizes[])
+void getConfigValue(int* length_alphabet, double* mu, double* maximal_distance, double* momentum, int* layerNb, int sizes[])
 {
 	ifstream	optionsFile(NAME_CONFIG_FILE);
 	string		line;
@@ -185,89 +185,9 @@ int* getConfigValue(int* length_alphabet, double* mu, double* maximal_distance, 
 				*layerNb = index+1;
 
 				// On recopie dans sizes
-				for (size_t j = 0; j < *layerNb; j++)
+				for (int j = 0; j < *layerNb; j++)
 					sizes[j] = sizesbis[j];
 			}
 		}
 	}
-}
-
-bool readExemple(char* nom_fichier, double entrees[], int taille_entree, string directory)
-{
-	char* temp = new char[MAX_LENGTH_NAME_FILE];
-	strncpy(temp, nom_fichier, MAX_LENGTH_NAME_FILE);
-	ifstream file((directory + string(temp)).c_str());                                                         //ouverture du fichier
-
-	if (!file)
-	{
-		//ERREUR -> message dans erreur.txt, affiché dans la console
-		err("Apprentissage - Bug 1 : Exemple " + directory + string(temp) + " non lu.",1);
-		delete temp;
-		return false;
-	}
-	else
-	{
-		// On lit le fichier
-		for (int i(0); i < taille_entree; i++)
-			file >> entrees[i];
-
-		delete temp;                                                                                                                                                                        // on insère les valeurs dans entrees
-		return true;
-	}
-}
-
-int countExemples(string directory)
-{
-	int nombre_fichiers(0);                                                         //Initialisation
-	DIR*   dp;
-	struct dirent* ep;
-
-	dp = opendir(directory.c_str());                                                         //On ouvre le répertoire contenant les images avant de le parcourir
-	if (dp != NULL)                                                         //Si l'ouverture du répertoire a fonctionné
-	{
-		while ((ep = readdir(dp)))                                                                                                                 //On parcourt le répertoire
-			if (strlen(ep->d_name) >= 4)                                                                                                                                                                         //Si le nom du fichier est d'au moins 4 caractères du genre .png ou .bmp (sinon le programme détecte d'autres fichiers invisibles bizarres)
-				nombre_fichiers++;
-		closedir(dp);                                                                                                                 //Fermeture du répertoire d'images
-	}
-	return nombre_fichiers;
-}
-
-int getArrayOfFileNames(char** tabloFichiers, string directory)
-{
-	int compteur(0);
-	DIR*   dp;
-	struct dirent* ep;
-
-	//On ouvre le répertoire contenant les images avant de le parcourir
-	dp = opendir(directory.c_str());
-
-	//Si l'ouverture du répertoire a fonctionné
-	if (dp != NULL)
-	{
-		//On parcourt le répertoire
-		while ((ep = readdir(dp)))
-		{
-			//Si le nom du fichier est d'au moins 4 caractères du genre .png ou .bmp (sinon le programme détecte d'autres fichiers invisibles bizarres)
-			if (strlen(ep->d_name) >= 4 && ep->d_name[0]!=0)
-			{
-				//on inscrit le nom du fichier dans le tablo
-				strncpy(tabloFichiers[compteur], ep->d_name,MAX_LENGTH_NAME_FILE);
-				compteur++;
-			}
-		}
-
-		//Fermeture du répertoire d'images
-		closedir(dp);
-	}
-	return compteur;
-}
-
-void getArrayOfExemples(char** tabloFichiers, double** tabloExemple, int nb_exemples, string directory)
-{
-	//cout << "Lecture des exemples ... " << flush;
-	for (int i(0); i < nb_exemples; i++)
-		//on lit chacun des exemples
-		readExemple(tabloFichiers[i], tabloExemple[i], FIRST_LAYER_SIZE, directory);
-	//cout << "Lecture terminee, " << countExemples(DOSSIERTEXTES) << " exemples lus." << endl;
 }
